backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0
yt-dlp>=2024.0.0
youtube-transcript-api==1.0.1
requests>=2.31.0
websocket-client>=1.7.0
